requirements.txt
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
flake8==3.8.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.18.4
pandas==1.0.3
pudb==2019.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
pyzillow==0.5.5
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
urwid==2.1.0
wrapt==1.12.1
zipp==3.1.0
